fix(scanner): correct cross-scanner dedup docs; refactor _deduplicate
The review-bot found a real doc/feature mismatch this PR introduced: the
docstring, an inline comment, and docs/cli/guard.md all newly claimed that on
the default (--no-skip-duplicate) path "the same secret reported at one location
by several scanners collapses to a single finding". That is false. The default
key includes rule_id, and each scanner namespaces its rule_ids (gitleaks-...,
trufflehog-..., native bare), so the same secret found by two scanners has
different keys and both survive. Cross-scanner same-secret collapse only happens
under --skip-duplicate (which keys on secret_hash alone).

- Correct the three statements to describe the real behavior.
- Add test_deduplicate_same_secret_two_scanners_not_collapsed_on_default_path,
  asserting both survive by default and collapse under skip_duplicate, so the
  docs cannot drift back to the false claim.
- Refactor _deduplicate: extract _dedup_key and a flat _should_replace (early
  returns) plus a _tie_key staticmethod, cutting the method's cyclomatic
  complexity instead of suppressing the CodeScene warning. Behavior-equivalent.
- Move the TestDeduplication class to tests/scanner/test_engine_dedup.py so
  neither module exceeds the functions-per-module threshold.
